+++ client/cpp/include/krpc/stream_impl.hpp
@@ -24,6 +24,8 @@ class StreamImpl {
   const std::string& get_data();
   bool is_null() const;
   void update(const std::string& data, bool is_null, const std::exception_ptr& exception);
+  void update_and_notify(const std::string& data, bool is_null,
+                         const std::exception_ptr& exception);
   bool has_updated() const;
   std::condition_variable& get_condition();
   std::unique_lock<std::mutex>& get_condition_lock();
@@ -40,6 +42,7 @@ class StreamImpl {
   std::recursive_mutex* update_lock;
   bool started;
   bool updated;
+  bool removed;
   std::string data;
   bool _is_null;
   std::exception_ptr exception;